group finders implementations under "-private"
