+++ addon-test-support/extend/find-element.js
@@ -1,11 +1,3 @@
-import $ from '-jquery';
-import {
-  buildSelector,
-  findClosestValue,
-  guardMultiple
-} from '../-private/helpers';
-import { getExecutionContext } from '../-private/execution_context';
-
 /**
  * @public
  *
@@ -39,19 +31,4 @@ import { getExecutionContext } from '../-private/execution_context';
  *
  * @throws Will throw an error if multiple elements are matched by selector and multiple option is not set
  */
-export function findElement(pageObjectNode, targetSelector, options = {}) {
-  const selector = buildSelector(pageObjectNode, targetSelector, options);
-  const container = getContainer(pageObjectNode, options);
-
-  let $elements = $(selector, container);
-
-  guardMultiple($elements, selector, options.multiple);
-
-  return $elements;
-}
-
-function getContainer(pageObjectNode, options) {
-  return options.testContainer
-    || findClosestValue(pageObjectNode, 'testContainer')
-    || getExecutionContext(pageObjectNode).testContainer;
-}
+export { findElement } from '../-private/finders';
